feat: changed srcipt.js
