+++ index.html
@@ -20,7 +20,7 @@
   <script src="episodes.js"></script>
  
   <!-- Loads YOUR javascript code -->
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
 </body>
 
 </html>
