+++ script.js
@@ -1,3 +1,4 @@
+import createEpisodeCard from "./episeodecard.js";
 //all Episodes
 let allEpisodes = [];
 //Setup
@@ -8,55 +9,55 @@ function setup() {
     createOptions(allEpisodes); //available episodes in the selection
 }
 
-function createEpisodeCard(allEpisodes) {
-    const episodeList = document.createElement('ul'); // Episode list
-    episodeList.classList.add('main'); //main class attached to list
-    allEpisodes.forEach((episode) => {
-        //for each episode a li card is created
-        const lI = document.createElement('li');
-        lI.className = 'episode-card';
+// function createEpisodeCard(allEpisodes) {
+//     const episodeList = document.createElement('ul'); // Episode list
+//     episodeList.classList.add('main'); //main class attached to list
+//     allEpisodes.forEach((episode) => {
+//         //for each episode a li card is created
+//         const lI = document.createElement('li');
+//         lI.className = 'episode-card';
 
-        const name = episode.name,         //individual episode data
-            season = episode.season,
-            number = episode.number,
-            summary = episode.summary.substring(3, episode.summary.length - 4); // removed summary tags that each episode had in summary
-        image = episode.image.medium;
+//         const name = episode.name,         //individual episode data
+//             season = episode.season,
+//             number = episode.number,
+//             summary = episode.summary.substring(3, episode.summary.length - 4); // removed summary tags that each episode had in summary
+//         image = episode.image.medium;
 
-        // const nameEl = document.createElement('h2'); //heading of episode /name of episode
-        // nameEl.className = 'episode-name';
+//         // const nameEl = document.createElement('h2'); //heading of episode /name of episode
+//         // nameEl.className = 'episode-name';
 
 
-        const formattedSeason = ('' + season).padStart(2, '0'); //type coersion
-        const formattedNumber = ('' + number).padStart(2, '0'); //type coersion
-        const episodeVersion = `S${formattedSeason}E${formattedNumber}`; //full version to present
+//         const formattedSeason = ('' + season).padStart(2, '0'); //type coersion
+//         const formattedNumber = ('' + number).padStart(2, '0'); //type coersion
+//         const episodeVersion = `S${formattedSeason}E${formattedNumber}`; //full version to present
 
-        const episodeVersionH2 = document.createElement('h2'); // episode heading
-        episodeVersionH2.innerText = name + ' - ' + episodeVersion;
+//         const episodeVersionH2 = document.createElement('h2'); // episode heading
+//         episodeVersionH2.innerText = name + ' - ' + episodeVersion;
 
-        const imageContainer = document.createElement('div'); // episode image
-        imageContainer.classList.add('image-container');
-        const episodeImg = document.createElement('img');
-        imageContainer.appendChild(episodeImg);
-        episodeImg.src = image;                        // image source is link to medium sized image
-        episodeImg.alt = 'show episode';
+//         const imageContainer = document.createElement('div'); // episode image
+//         imageContainer.classList.add('image-container');
+//         const episodeImg = document.createElement('img');
+//         imageContainer.appendChild(episodeImg);
+//         episodeImg.src = image;                        // image source is link to medium sized image
+//         episodeImg.alt = 'show episode';
 
-        const episodeSummary = document.createElement('p');  //episode summary
-        episodeSummary.classList.add('summary');
-        episodeSummary.innerText = summary;
+//         const episodeSummary = document.createElement('p');  //episode summary
+//         episodeSummary.classList.add('summary');
+//         episodeSummary.innerText = summary;
 
-        lI.appendChild(episodeVersionH2);
-        lI.appendChild(imageContainer);
-        lI.appendChild(episodeSummary);
+//         lI.appendChild(episodeVersionH2);
+//         lI.appendChild(imageContainer);
+//         lI.appendChild(episodeSummary);
 
-        episodeList.appendChild(lI); // appended li to ul element
-    });
+//         episodeList.appendChild(lI); // appended li to ul element
+//     });
 
-    const rootEl = document.getElementById('root');
+//     const rootEl = document.getElementById('root');
 
-    // targeting root element in the DOM
+//     // targeting root element in the DOM
 
-    rootEl.appendChild(episodeList);
-}
+//     rootEl.appendChild(episodeList);
+// }
 
 // header body and footer styling and data presentation 
 function makePageForEpisodes() {
